Update FantasyFootball.py
Added in commented out code as well as deleted outdated commented out code that is replaced by more efficient lines. Figured out how to iterate through and create url's to match rushing, receiving, and passing stats. Created a couple of functions to simplify getting the rows, and columns from the scraped websites. Defined a soupify function that will automatically open the url and beautifuly soupify it.